requirements.txt
faiss-cpu
sentence-transformers
langchain
torch

